merge new changes from base

+++ src/scripts/ui/ux/tooltip.ts
@@ -56,6 +56,13 @@ function createTooltipInstanceIfMissing(target: Element) {
     undefined) as Placement | undefined;
   const distance = +target.getAttribute("data-tooltip-distance")! || 0;
 
+  let maxWidth: number | undefined = undefined;
+  if (target.getAttribute("data-tooltip-wide") !== null) {
+    maxWidth = 860;
+  } else if (target.getAttribute("data-tooltip-med") !== null) {
+    maxWidth = 480;
+  }
+
   instance = tippy(target, {
     content: content_element,
     duration: [showAnimDuration, 0], // animation duration
@@ -63,8 +70,7 @@ function createTooltipInstanceIfMissing(target: Element) {
       target.getAttribute("data-tooltip-delay") !== null
         ? [500, 0]
         : [showDelay, 0], // time on hover before tooltip opens
-    maxWidth:
-      target.getAttribute("data-tooltip-wide") !== null ? 860 : undefined,
+    maxWidth,
     trigger:
       target.getAttribute("data-tooltip-noclick") !== null
         ? "mouseenter"
@@ -106,7 +112,7 @@ function findDataTooltip(ev: Event): Element | null {
   return null;
 }
 
-let checkTooltipReferencesTimeout: ReturnType<typeof setTimeout> | null = null; // to debounce calls to checkTooltipReferences
+let checkTooltipReferencesQueued = false;
 
 function destroyTooltip(instance: TooltipInstance) {
   instance.setProps({
@@ -120,12 +126,13 @@ function destroyTooltip(instance: TooltipInstance) {
 // Checks that all tooltip references (targets) are still valid DOM elements
 // attached to document, otherwise destroy those instances
 function checkTooltipReferences() {
-  checkTooltipReferencesTimeout = null;
+  checkTooltipReferencesQueued = false;
 
-  for (const instance of tooltip_instances.filter(
-    (instance) => !instance.reference || !instance.reference.isConnected,
-  )) {
-    destroyTooltip(instance);
+  for (let i = 0; i < tooltip_instances.length; ++i) {
+    const instance = tooltip_instances[i];
+    if (!instance.reference || !instance.reference.isConnected) {
+      destroyTooltip(instance);
+    }
   }
 }
 
@@ -151,9 +158,10 @@ function setupMutationObserver() {
   setTimeout(function () {
     const observer = new MutationObserver(
       function (/*mutationsList, observer*/) {
-        if (!checkTooltipReferencesTimeout)
-          // call checkTooltipReferences, debounced
-          checkTooltipReferencesTimeout = setTimeout(checkTooltipReferences, 1);
+        if (!checkTooltipReferencesQueued) {
+          checkTooltipReferencesQueued = true;
+          queueMicrotask(checkTooltipReferences);
+        }
       },
     );
